libs/service: Move into internal
