+++ abci/client/client.go
@@ -5,7 +5,7 @@ import (
 	"sync"
 
 	"github.com/cometbft/cometbft/abci/types"
-	"github.com/cometbft/cometbft/libs/service"
+	"github.com/cometbft/cometbft/internal/service"
 	cmtsync "github.com/cometbft/cometbft/libs/sync"
 )
