+++ abci/client/grpc_client.go
@@ -12,7 +12,7 @@ import (
 
 	"github.com/cometbft/cometbft/abci/types"
 	cmtnet "github.com/cometbft/cometbft/internal/net"
-	"github.com/cometbft/cometbft/libs/service"
+	"github.com/cometbft/cometbft/internal/service"
 )
 
 var _ Client = (*grpcClient)(nil)
