+++ abci/client/local_client.go
@@ -4,7 +4,7 @@ import (
 	"context"
 
 	types "github.com/cometbft/cometbft/abci/types"
-	"github.com/cometbft/cometbft/libs/service"
+	"github.com/cometbft/cometbft/internal/service"
 	cmtsync "github.com/cometbft/cometbft/libs/sync"
 )
