+++ abci/server/server.go
@@ -9,7 +9,7 @@ package server
 
 import (
 	"github.com/cometbft/cometbft/abci/types"
-	"github.com/cometbft/cometbft/libs/service"
+	"github.com/cometbft/cometbft/internal/service"
 )
 
 // NewServer is a utility function for out of process applications to set up either a socket or
