+++ config/db.go
@@ -5,8 +5,8 @@ import (
 
 	dbm "github.com/cometbft/cometbft-db"
 
+	"github.com/cometbft/cometbft/internal/service"
 	"github.com/cometbft/cometbft/libs/log"
-	"github.com/cometbft/cometbft/libs/service"
 )
 
 // ServiceProvider takes a config and a logger and returns a ready to go Node.
