+++ internal/blocksync/pool.go
@@ -8,8 +8,8 @@ import (
 	"time"
 
 	flow "github.com/cometbft/cometbft/internal/flowrate"
+	"github.com/cometbft/cometbft/internal/service"
 	"github.com/cometbft/cometbft/libs/log"
-	"github.com/cometbft/cometbft/libs/service"
 	cmtsync "github.com/cometbft/cometbft/libs/sync"
 	"github.com/cometbft/cometbft/p2p"
 	"github.com/cometbft/cometbft/types"
