+++ internal/consensus/ticker.go
@@ -3,13 +3,11 @@ package consensus
 import (
 	"time"
 
+	"github.com/cometbft/cometbft/internal/service"
 	"github.com/cometbft/cometbft/libs/log"
-	"github.com/cometbft/cometbft/libs/service"
 )
 
-var (
-	tickTockBufferSize = 10
-)
+var tickTockBufferSize = 10
 
 // TimeoutTicker is a timer that schedules timeouts
 // conditional on the height/round/step in the timeoutInfo.
@@ -50,7 +48,6 @@ func NewTimeoutTicker() TimeoutTicker {
 
 // OnStart implements service.Service. It starts the timeout routine.
 func (t *timeoutTicker) OnStart() error {
-
 	go t.timeoutRoutine()
 
 	return nil
