+++ internal/events/events.go
@@ -4,7 +4,7 @@ package events
 import (
 	"fmt"
 
-	"github.com/cometbft/cometbft/libs/service"
+	"github.com/cometbft/cometbft/internal/service"
 	cmtsync "github.com/cometbft/cometbft/libs/sync"
 )
