+++ internal/pubsub/pubsub.go
@@ -38,7 +38,7 @@ import (
 	"errors"
 	"fmt"
 
-	"github.com/cometbft/cometbft/libs/service"
+	"github.com/cometbft/cometbft/internal/service"
 	cmtsync "github.com/cometbft/cometbft/libs/sync"
 )
 
@@ -150,7 +150,8 @@ func (s *Server) Subscribe(
 	ctx context.Context,
 	clientID string,
 	query Query,
-	outCapacity ...int) (*Subscription, error) {
+	outCapacity ...int,
+) (*Subscription, error) {
 	outCap := 1
 	if len(outCapacity) > 0 {
 		if outCapacity[0] <= 0 {
