+++ internal/state/pruner.go
@@ -6,10 +6,10 @@ import (
 	"time"
 
 	"github.com/cometbft/cometbft/config"
+	"github.com/cometbft/cometbft/internal/service"
 	"github.com/cometbft/cometbft/internal/state/indexer"
 	"github.com/cometbft/cometbft/internal/state/txindex"
 	"github.com/cometbft/cometbft/libs/log"
-	"github.com/cometbft/cometbft/libs/service"
 )
 
 var (
