+++ light/rpc/client.go
@@ -9,10 +9,10 @@ import (
 	"time"
 
 	"github.com/cometbft/cometbft/crypto/merkle"
+	service "github.com/cometbft/cometbft/internal/service"
 	"github.com/cometbft/cometbft/internal/state"
 	cmtbytes "github.com/cometbft/cometbft/libs/bytes"
 	cmtmath "github.com/cometbft/cometbft/libs/math"
-	service "github.com/cometbft/cometbft/libs/service"
 	rpcclient "github.com/cometbft/cometbft/rpc/client"
 	ctypes "github.com/cometbft/cometbft/rpc/core/types"
 	rpctypes "github.com/cometbft/cometbft/rpc/jsonrpc/types"
