+++ p2p/peer.go
@@ -9,8 +9,8 @@ import (
 	"github.com/cosmos/gogoproto/proto"
 
 	"github.com/cometbft/cometbft/internal/cmap"
+	"github.com/cometbft/cometbft/internal/service"
 	"github.com/cometbft/cometbft/libs/log"
-	"github.com/cometbft/cometbft/libs/service"
 
 	cmtconn "github.com/cometbft/cometbft/p2p/conn"
 )
