+++ proxy/multi_app_conn.go
@@ -5,8 +5,8 @@ import (
 
 	abcicli "github.com/cometbft/cometbft/abci/client"
 	cmtos "github.com/cometbft/cometbft/internal/os"
+	"github.com/cometbft/cometbft/internal/service"
 	cmtlog "github.com/cometbft/cometbft/libs/log"
-	"github.com/cometbft/cometbft/libs/service"
 )
 
 const (
