+++ rpc/jsonrpc/client/ws_client.go
@@ -13,8 +13,8 @@ import (
 	metrics "github.com/rcrowley/go-metrics"
 
 	cmtrand "github.com/cometbft/cometbft/internal/rand"
+	"github.com/cometbft/cometbft/internal/service"
 	"github.com/cometbft/cometbft/libs/log"
-	"github.com/cometbft/cometbft/libs/service"
 	cmtsync "github.com/cometbft/cometbft/libs/sync"
 	types "github.com/cometbft/cometbft/rpc/jsonrpc/types"
 )
