+++ rpc/jsonrpc/server/ws_handler.go
@@ -12,8 +12,8 @@ import (
 
 	"github.com/gorilla/websocket"
 
+	"github.com/cometbft/cometbft/internal/service"
 	"github.com/cometbft/cometbft/libs/log"
-	"github.com/cometbft/cometbft/libs/service"
 	types "github.com/cometbft/cometbft/rpc/jsonrpc/types"
 )
