+++ types/event_bus.go
@@ -6,8 +6,8 @@ import (
 
 	"github.com/cometbft/cometbft/abci/types"
 	cmtpubsub "github.com/cometbft/cometbft/internal/pubsub"
+	"github.com/cometbft/cometbft/internal/service"
 	"github.com/cometbft/cometbft/libs/log"
-	"github.com/cometbft/cometbft/libs/service"
 )
 
 const defaultCapacity = 0
